FINALLY the code font size should be fixed
Apparently the problem was a mixture of both the font-size AND the font of one of them being overwritten.

After fixing them it also became clear that for some reason their was a noticeable shift between the two texts so that they never align - I pinpoited the issue to mainly being one had a border and the other didn't, so I added what should be an invisible border and added a tiny shift from the left to make the text appear almost perfectly aligned.

Despite all this, the issue is still not totally fixed - the width is still way to variable - so that sometimes part of the text will clip to the next line after running out of space while the other text still fits - making the rest of the thing unaligned. Thankfully the difference in width appears to be no more than 2-4 px at best and it's unnoticeable most of the time - but I'll work on fixing it later.