Even more code lines now (fix part 2)
I think I've fixed the width issue - however for some reason beyond human comprehension it appears impossible to 100% align the two texts. So for now I've adjusted it just enough that it is feasibly impossible (outside zoom 250% and above... which I doubt many people are using) for the line cursor to deviate away from the letter it's actually at. I'm going to call it a fix for now until I can find a better way to handle it.